added new bonus files and fixed some other files, renamed everything with ft_

+++ ft_intlength.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-int	intlength(int n)
+int	intlength(long n)
 {
 	int	i;
